suggested edits necessary to make a python module
