+++ ccx2paraview/clean.py
